Split constraints module into more files
