+++ pyproject.toml
@@ -114,4 +114,4 @@ exclude = [
 ]
 
 [tool.pytest.ini_options]
-addopts = "-n auto --timeout 700"
+addopts = "-n auto --timeout 700 --duration=10"
