+++ src/fandango/api.py
@@ -4,7 +4,8 @@ import itertools
 import logging
 import time
 from typing import IO, Optional
-from fandango.constraints.base import Constraint, SoftValue
+from fandango.constraints.constraint import Constraint
+from fandango.constraints.soft import SoftValue
 from fandango.language.grammar import FuzzingMode, ParsingMode
 from fandango.language.grammar.grammar import Grammar
 from fandango.language.parse import parse
