+++ src/fandango/cli/__init__.py
@@ -10,7 +10,8 @@ import re
 from typing import IO, Any
 from collections.abc import Callable
 
-from fandango.constraints.base import Constraint, SoftValue
+from fandango.constraints.soft import SoftValue
+from fandango.constraints.constraint import Constraint
 from fandango.converters.FandangoConverter import FandangoConverter
 from fandango.language.tree import DerivationTree
