+++ src/fandango/constraints/__init__.py
@@ -1,0 +1,5 @@
+"""
+This module contains the base classes for constraints in the fandango library.
+"""
+
+LEGACY = False
