+++ src/fandango/evolution/algorithm.py
@@ -8,7 +8,8 @@ import warnings
 from collections.abc import Callable, Generator
 from typing import Iterable, Optional, Union
 
-from fandango.constraints.base import Constraint, SoftValue
+from fandango.constraints.constraint import Constraint
+from fandango.constraints.soft import SoftValue
 from fandango.errors import FandangoFailedError, FandangoParseError, FandangoValueError
 from fandango.evolution import GeneratorWithReturn
 from fandango.evolution.adaptation import AdaptiveTuner
