+++ src/fandango/language/convert.py
@@ -3,19 +3,17 @@ from collections.abc import Sequence
 from io import UnsupportedOperation
 from typing import Any, Optional
 
+from fandango.constraints.soft import SoftValue
+from fandango.constraints.repetition_bounds import RepetitionBoundsConstraint
 from fandango.errors import FandangoValueError
-from fandango.constraints.base import (
-    Constraint,
-    ComparisonConstraint,
-    ConjunctionConstraint,
-    DisjunctionConstraint,
-    ExistsConstraint,
-    ExpressionConstraint,
-    ForallConstraint,
-    SoftValue,
-    RepetitionBoundsConstraint,
-)
-from fandango.constraints.fitness import Comparison
+from fandango.constraints.constraint import Constraint
+from fandango.constraints.comparison import ComparisonConstraint
+from fandango.constraints.conjunction import ConjunctionConstraint
+from fandango.constraints.disjunct import DisjunctionConstraint
+from fandango.constraints.exists import ExistsConstraint
+from fandango.constraints.expression import ExpressionConstraint
+from fandango.constraints.forall import ForallConstraint
+from fandango.constraints.failing_tree import Comparison
 from fandango.language import NonTerminalSearch, NodeType
 from fandango.language.grammar import FuzzingMode
 from fandango.language.grammar.grammar import Grammar
